backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
sentence-transformers==2.7.0
chromadb==0.5.0
google-generativeai==0.5.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30